requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
alembic==1.18.4
anyio==4.12.1
apache-airflow==2.9.3
apache-airflow-providers-common-compat==1.6.1
apache-airflow-providers-common-io==1.5.4
apache-airflow-providers-common-sql==1.26.0
apache-airflow-providers-fab==1.5.3
apache-airflow-providers-ftp==3.12.3
apache-airflow-providers-http==5.2.2
apache-airflow-providers-imap==3.8.3
apache-airflow-providers-smtp==2.0.3
apache-airflow-providers-sqlite==4.0.2
apispec==6.10.0
argcomplete==3.6.3
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asgiref==3.11.1
asttokens==3.0.1
async-lru==2.2.0
attrs==25.4.0
babel==2.18.0
beautifulsoup4==4.14.3
bleach==6.3.0
blinker==1.9.0
cachelib==0.13.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
comm==0.2.3
ConfigUpdater==3.2
connexion==2.14.2
cron_descriptor==2.0.8
croniter==6.2.2
cryptography==46.0.6
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.3.1
dill==0.4.1
dnspython==2.8.0
email-validator==2.3.0
executing==2.2.1
fastjsonschema==2.21.2
Flask==2.2.5
Flask-AppBuilder==4.5.3
Flask-Babel==2.0.0
Flask-Caching==2.3.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2026.3.0
google-re2==1.1.20251105
googleapis-common-protos==1.73.1
greenlet==3.3.2
grpcio==1.80.0
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
inflection==0.5.1
ipykernel==7.2.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.1.0
json5==0.13.0
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
lark==1.3.1
lazy-object-proxy==1.12.0
limits==5.8.0
linkify-it-py==2.1.0
lockfile==0.12.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
marshmallow-oneofschema==3.2.0
marshmallow-sqlalchemy==0.28.2
matplotlib-inline==0.2.1
mdit-py-plugins==0.5.0
mdurl==0.1.2
methodtools==0.4.7
mistune==3.2.0
more-itertools==10.8.0
multidict==6.7.1
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.5.5
notebook_shim==0.2.4
numpy==2.4.3
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-exporter-otlp-proto-http==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
ordered-set==4.1.0
packaging==26.0
pandas==3.0.1
pandocfilters==1.5.1
parso==0.8.6
pathlib_abc==0.5.2
pathspec==1.0.4
pendulum==3.2.0
pexpect==4.9.0
platformdirs==4.9.4
pluggy==1.6.0
prison==0.2.1
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.6
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.9
pycparser==3.0
Pygments==2.19.2
PyJWT==2.12.1
pyspark==4.1.1
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2026.1.post1
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==13.9.4
rich-argparse==1.7.2
rpds-py==0.30.0
Send2Trash==2.1.0
setproctitle==1.3.7
setuptools==82.0.1
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.42.1
sqlparse==0.5.5
stack-data==0.6.3
tabulate==0.10.0
tenacity==9.1.4
termcolor==3.3.0
terminado==0.18.1
text-unidecode==1.3
tinycss2==1.4.0
tornado==6.5.5
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.3
uc-micro-py==2.0.0
unicodecsv==0.14.1
universal_pathlib==0.3.10
uri-template==1.3.0
urllib3==2.6.3
wcwidth==0.6.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==2.2.3
wirerope==1.0.0
wrapt==2.1.2
WTForms==3.2.1
yarl==1.23.0
zipp==3.23.0
